// src/app/models/sms/GetMaquinasModel.ts
export class GetMaquinasModel {
    idMaquina: string = '';
    nome_maquina: string = '';
    local_maquina: string = '';
    descricao_maquina: string = ''; // Use `descricao_maquina` consistentemente
    descricao_funcao_maquina: string = '';
    patrimonio_maquina: string = '';
    fabricante_maquina: string = '';
    fabricacao_maquina: string = '';
    serie_maquina: string = '';
    modelo_maquina: string = '';
    posicaotrabalho_maquina: string = '';
    peso_maquina: string = '';
    historico_acidente_maquina: string = '';
    voltagem_maquina: string = '';
    comando_maquina: string = '';
    potencia_maquina: string = '';
    qtdOperadores_maquina: string = '';
    turno_maquina: string = '';
    dias_turno_maquina: string = '';
    capacidade_maquina: string = '';
    carga_maquina: string = '';
    descarga_maquina: string = '';
    arranjo_fisico: string = '';
    instalaçoes: string = '';
    dispositivos_eletricos: string = '';
    aterramento_maquinas: string = '';
    condutores_maquinas: string = '';
    dispositivos_partida: string = '';
    sistema_seguranca: string = '';
  }
  